+++ src/actions/failed-check.ts
@@ -1,13 +1,13 @@
 import type { CoderClient } from "../services/coder/client";
 import type { GitHubClient, PRInfo } from "../services/github/client";
 import type { Logger } from "../infra/logger";
-import { MAX_FAILED_JOBS, formatFailedCheckMessage } from "../messages";
+import { MAX_FAILED_JOBS, formatFailedCheckMessage } from "./messages";
 import type { ActionOutputs, HandlerConfig } from "../config/handler-config";
 import {
 	generateTaskName,
 	lookupAndEnsureActiveTask,
 	sendInputWithRetry,
-} from "../task-utils";
+} from "./task-utils";
 
 const MAX_LOG_LINES = 100;
