+++ src/actions/pr-comment.ts
@@ -1,13 +1,13 @@
 import type { CoderClient } from "../services/coder/client";
 import type { GitHubClient } from "../services/github/client";
 import type { Logger } from "../infra/logger";
-import { formatPRCommentMessage } from "../messages";
+import { formatPRCommentMessage } from "./messages";
 import type { ActionOutputs, HandlerConfig } from "../config/handler-config";
 import {
 	generateTaskName,
 	lookupAndEnsureActiveTask,
 	sendInputWithRetry,
-} from "../task-utils";
+} from "./task-utils";
 
 export interface PRCommentContext {
 	owner: string;
