+++ src/actions/task-utils.test.ts
@@ -1,8 +1,8 @@
 import { beforeEach, describe, expect, mock, test } from "bun:test";
-import type { CoderClient } from "./services/coder/client";
-import { CoderAPIError } from "./services/coder/client";
-import { TestLogger } from "./infra/logger";
-import { MockCoderClient, mockTask } from "./testing/helpers";
+import type { CoderClient } from "../services/coder/client";
+import { CoderAPIError } from "../services/coder/client";
+import { TestLogger } from "../infra/logger";
+import { MockCoderClient, mockTask } from "../testing/helpers";
 import {
 	generateTaskName,
 	lookupAndEnsureActiveTask,
