+++ src/events/dispatcher.ts
@@ -1,10 +1,10 @@
 import type { CoderClient } from "../services/coder/client";
 import { GitHubClient, type Octokit } from "../services/github/client";
-import { CloseTaskHandler } from "../handlers/close-task";
-import { CreateTaskHandler } from "../handlers/create-task";
-import { FailedCheckHandler } from "../handlers/failed-check";
-import { IssueCommentHandler } from "../handlers/issue-comment";
-import { PRCommentHandler } from "../handlers/pr-comment";
+import { CloseTaskHandler } from "../actions/close-task";
+import { CreateTaskHandler } from "../actions/create-task";
+import { FailedCheckHandler } from "../actions/failed-check";
+import { IssueCommentHandler } from "../actions/issue-comment";
+import { PRCommentHandler } from "../actions/pr-comment";
 import type { Logger } from "../infra/logger";
 import type { ActionOutputs, HandlerConfig } from "../config/handler-config";
 import type { AppConfig } from "../config/app-config";
